fix(dashboards): part 1 of refactoring
