+++ pkg/query-service/app/cloudintegrations/controller.go
@@ -11,6 +11,7 @@ import (
 	"github.com/SigNoz/signoz/pkg/query-service/model"
 	"github.com/SigNoz/signoz/pkg/sqlstore"
 	"github.com/SigNoz/signoz/pkg/types"
+	"github.com/SigNoz/signoz/pkg/types/dashboardtypes"
 	"golang.org/x/exp/maps"
 )
 
@@ -510,10 +511,8 @@ func (c *Controller) UpdateServiceConfig(
 
 // All dashboards that are available based on cloud integrations configuration
 // across all cloud providers
-func (c *Controller) AvailableDashboards(ctx context.Context, orgId string) (
-	[]types.Dashboard, *model.ApiError,
-) {
-	allDashboards := []types.Dashboard{}
+func (c *Controller) AvailableDashboards(ctx context.Context, orgId string) ([]*dashboardtypes.Dashboard, *model.ApiError) {
+	allDashboards := []*dashboardtypes.Dashboard{}
 
 	for _, provider := range []string{"aws"} {
 		providerDashboards, apiErr := c.AvailableDashboardsForCloudProvider(ctx, orgId, provider)
@@ -529,10 +528,7 @@ func (c *Controller) AvailableDashboards(ctx context.Context, orgId string) (
 	return allDashboards, nil
 }
 
-func (c *Controller) AvailableDashboardsForCloudProvider(
-	ctx context.Context, orgID string, cloudProvider string,
-) ([]types.Dashboard, *model.ApiError) {
-
+func (c *Controller) AvailableDashboardsForCloudProvider(ctx context.Context, orgID string, cloudProvider string) ([]*dashboardtypes.Dashboard, *model.ApiError) {
 	accountRecords, apiErr := c.accountsRepo.listConnected(ctx, orgID, cloudProvider)
 	if apiErr != nil {
 		return nil, model.WrapApiError(apiErr, "couldn't list connected cloud accounts")
@@ -563,16 +559,15 @@ func (c *Controller) AvailableDashboardsForCloudProvider(
 		return nil, apiErr
 	}
 
-	svcDashboards := []types.Dashboard{}
+	svcDashboards := []*dashboardtypes.Dashboard{}
 	for _, svc := range allServices {
 		serviceDashboardsCreatedAt := servicesWithAvailableMetrics[svc.Id]
 		if serviceDashboardsCreatedAt != nil {
 			for _, d := range svc.Assets.Dashboards {
-				isLocked := 1
 				author := fmt.Sprintf("%s-integration", cloudProvider)
-				svcDashboards = append(svcDashboards, types.Dashboard{
-					UUID:   c.dashboardUuid(cloudProvider, svc.Id, d.Id),
-					Locked: &isLocked,
+				svcDashboards = append(svcDashboards, &dashboardtypes.Dashboard{
+					ID:     c.dashboardUuid(cloudProvider, svc.Id, d.Id),
+					Locked: true,
 					Data:   *d.Definition,
 					TimeAuditable: types.TimeAuditable{
 						CreatedAt: *serviceDashboardsCreatedAt,
@@ -590,11 +585,7 @@ func (c *Controller) AvailableDashboardsForCloudProvider(
 
 	return svcDashboards, nil
 }
-func (c *Controller) GetDashboardById(
-	ctx context.Context,
-	orgId string,
-	dashboardUuid string,
-) (*types.Dashboard, *model.ApiError) {
+func (c *Controller) GetDashboardById(ctx context.Context, orgId string, dashboardUuid string) (*dashboardtypes.Dashboard, *model.ApiError) {
 	cloudProvider, _, _, apiErr := c.parseDashboardUuid(dashboardUuid)
 	if apiErr != nil {
 		return nil, apiErr
@@ -608,8 +599,8 @@ func (c *Controller) GetDashboardById(
 	}
 
 	for _, d := range allDashboards {
-		if d.UUID == dashboardUuid {
-			return &d, nil
+		if d.ID == dashboardUuid {
+			return d, nil
 		}
 	}
