+++ pkg/types/ttl.go
@@ -1,0 +1,17 @@
+package types
+
+import (
+	"github.com/uptrace/bun"
+)
+
+type TTLSetting struct {
+	bun.BaseModel `bun:"table:ttl_setting"`
+	Identifiable
+	TimeAuditable
+	TransactionID  string `bun:"transaction_id,type:text,notnull"`
+	TableName      string `bun:"table_name,type:text,notnull"`
+	TTL            int    `bun:"ttl,notnull,default:0"`
+	ColdStorageTTL int    `bun:"cold_storage_ttl,notnull,default:0"`
+	Status         string `bun:"status,type:text,notnull"`
+	OrgID          string `json:"-" bun:"org_id,notnull"`
+}
